skills/site-patrol/pyproject.toml
[project]
name = "site-patrol"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "feedparser>=6.0",
    "pyyaml>=6.0",
]

[tool.uv]
dev-dependencies = []
